+++ lib/widgets/simple_container/models/transform.dart
@@ -23,7 +23,9 @@ class STransform {
           tween:
               Tween<double>(begin: degrees * pi / 180, end: degrees * pi / 180),
           child: child,
-          onEnd: () => _style.animation.onEnd(SAnimationEnd.rotate),
+          onEnd: _style.animation.onEnd == null
+              ? null
+              : () => _style.animation.onEnd(SAnimationEnd.rotate),
           builder: (_, value, Widget child) {
             return Transform.rotate(angle: value, child: child);
           },
@@ -40,7 +42,9 @@ class STransform {
         return _style.animation._buildAnimatedTween(
           tween: Tween<double>(begin: scale, end: scale),
           child: child,
-          onEnd: () => _style.animation.onEnd(SAnimationEnd.scale),
+          onEnd: _style.animation.onEnd == null
+              ? null
+              : () => _style.animation.onEnd(SAnimationEnd.scale),
           builder: (_, value, Widget child) {
             return Transform.scale(scale: value, child: child);
           },
@@ -57,12 +61,16 @@ class STransform {
         return _style.animation._buildAnimatedTween(
           tween: Tween<double>(begin: dx ?? 0, end: dx ?? 0),
           child: child,
-          onEnd: () => _style.animation.onEnd(SAnimationEnd.translate),
+          onEnd: _style.animation.onEnd == null
+              ? null
+              : () => _style.animation.onEnd(SAnimationEnd.translate),
           builder: (_, dxValue, Widget child) {
             return _style.animation._buildAnimatedTween(
               tween: Tween<double>(begin: dy ?? 0, end: dy ?? 0),
               child: child,
-              onEnd: () => _style.animation.onEnd(SAnimationEnd.translate),
+              onEnd: _style.animation.onEnd == null
+                  ? null
+                  : () => _style.animation.onEnd(SAnimationEnd.translate),
               builder: (_, dyValue, Widget child) {
                 return Transform.translate(
                   offset: Offset(dxValue, dyValue),